android: java warning fix

+++ trunk/android/android/src/NativeInterface.java
@@ -119,11 +119,11 @@ public class NativeInterface
 		// hardcoded exceptions for known devices that return wrong DPI
 		if (Build.MANUFACTURER.equals("HTC") && Build.MODEL.equals("HTC One X"))
 		{
-			return 312.0;
+			return 312.0f;
 		}
 		DisplayMetrics metrics = new DisplayMetrics();
 		NativeInterface.Activity.getWindowManager().getDefaultDisplay().getMetrics(metrics);
-		return Math.sqrt((metrics.xdpi * metrics.xdpi + metrics.ydpi * metrics.ydpi) / 2.0);
+		return (float) Math.sqrt((metrics.xdpi * metrics.xdpi + metrics.ydpi * metrics.ydpi) / 2.0);
 	}
 	
 	public static String getOsVersion()
